requirements.txt
requests==2.28.1
kafka-python==2.0.2
pyspark==3.4.1
textblob==0.17.1
